fixed some statistics functions

+++ test/statistics.cpp
@@ -49,6 +49,6 @@ TEST_CASE("median", "median") {
     vector<int> c3 = { 2, 1, 5, 4, 3, 7, 9, 8, 6 };
     REQUIRE(median_c(c3) == 5);
 
-    vector<int> c4 = { 1, 2, 3, 4, 5.5, 6, 7, 8, 9, 10 };
+    vector<double> c4 = { 1, 2, 3, 4, 5.5, 6, 7, 8, 9, 10 };
     REQUIRE(median_c(c4) == 5.75);
 }
